// src/trading/models/Order.ts
export interface Order extends Record<string, unknown> {
    /**
     * 
     * @type {AssetClass}
     * @memberof Order
     */
    assetClass?: AssetClass;
    /**
     * Asset ID (For options this represents the option contract ID)
     * @type {string}
     * @memberof Order
     */
    assetId?: string;
    /**
     * 
     * @type {Date}
     * @memberof Order
     */
    canceledAt?: Date | null;
    /**
     * Client unique order ID
     * @type {string}
     * @memberof Order
     */
    clientOrderId?: string;
    /**
     * 
     * @type {Date}
     * @memberof Order
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Order
     */
    expiredAt?: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof Order
     */
    expiresAt?: Date;
    /**
     * If true, eligible for execution outside regular trading hours.
     * @type {boolean}
     * @memberof Order
     */
    extendedHours?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof Order
     */
    failedAt?: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof Order
     */
    filledAt?: Date | null;
    /**
     * Filled average price
     * @type {string}
     * @memberof Order
     */
    filledAvgPrice?: string | null;
    /**
     * Filled quantity
     * @type {string}
     * @memberof Order
     */
    filledQty?: string;
    /**
     * The highest (lowest) market price seen since the trailing stop order was submitted.
     * @type {string}
     * @memberof Order
     */
    hwm?: string | null;
    /**
     * Order ID
     * @type {string}
     * @memberof Order
     */
    id?: string;
    /**
     * When querying non-simple order_class orders in a nested style, an array of Order entities associated with this order. Otherwise, null. Required if order class is `mleg`.
     * @type {Array<OrderLeg>}
     * @memberof Order
     */
    legs?: Array<OrderLeg> | null;
    /**
     * Limit price
     * @type {string}
     * @memberof Order
     */
    limitPrice?: string | null;
    /**
     * Ordered notional amount. If entered, qty will be null. Can take up to 9 decimal points.
     * @type {string}
     * @memberof Order
     */
    notional: string | null;
    /**
     * 
     * @type {OrderClass}
     * @memberof Order
     */
    orderClass?: OrderClass;
    /**
     * Deprecated in favour of the field "type" 
     * @type {string}
     * @memberof Order
     * @deprecated
     */
    orderType?: string;
    /**
     * 
     * @type {PositionIntent}
     * @memberof Order
     */
    positionIntent?: PositionIntent;
    /**
     * Ordered quantity. If entered, notional will be null. Can take up to 9 decimal points. Required if order class is `mleg`.
     * @type {string}
     * @memberof Order
     */
    qty?: string | null;
    /**
     * The proportional quantity of this leg in relation to the overall multi-leg order quantity.
     * @type {string}
     * @memberof Order
     */
    ratioQty?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof Order
     */
    replacedAt?: Date | null;
    /**
     * The order ID that this order was replaced by
     * @type {string}
     * @memberof Order
     */
    replacedBy?: string | null;
    /**
     * The order ID that this order replaces
     * @type {string}
     * @memberof Order
     */
    replaces?: string | null;
    /**
     * 
     * @type {OrderSide}
     * @memberof Order
     */
    side?: OrderSide;
    /**
     * 
     * @type {OrderStatus}
     * @memberof Order
     */
    status?: OrderStatus;
    /**
     * Stop price
     * @type {string}
     * @memberof Order
     */
    stopPrice?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof Order
     */
    submittedAt?: Date | null;
    /**
     * Asset symbol, required for all order classes except for `mleg`
     * @type {string}
     * @memberof Order
     */
    symbol?: string;
    /**
     * 
     * @type {TimeInForce}
     * @memberof Order
     */
    timeInForce: TimeInForce;
    /**
     * The percent value away from the high water mark for trailing stop orders.
     * @type {string}
     * @memberof Order
     */
    trailPercent?: string | null;
    /**
     * The dollar value away from the high water mark for trailing stop orders.
     * @type {string}
     * @memberof Order
     */
    trailPrice?: string | null;
    /**
     * 
     * @type {OrderType}
     * @memberof Order
     */
    type: OrderType;
    /**
     * 
     * @type {Date}
     * @memberof Order
     */
    updatedAt?: Date | null;
}